Changed File Naming and added a Python utility
Changed file naming as GLSeq will likely become more heterozygous in the
nature of scripts I put in it.

The first python script I've added is rhFnaConverter, which takes an
input file with many chromosomes in one file and outputs a folder with a
containing file for each chromosome.  This is the format Rockhopper
uses, so it also outputs the part of the command that Rockhopper will
end up using.