Completed Condor enabled GLSeq Wrapper
This python program now interfaces with pydagman to construct, group,
and run GLSeq2 pipelines.  Of note, the GLSeq2 update that allows for
this has note been pushed yet as it is still under testing, but
preliminary tests indicate that running this wrapper's PyGLSeqWrapper
with the normal arguments (Replacing Rscript with python
PyGLSeqWrapper.py) will result in a condor enabled run.