Condor Functionality
Implemented more of the dataprep.  The graphs looks a lot nicer now that
I've switch to pygraphviz.  Also a bunch of optimizations both in the
wrapper and in the files to make things run faster/knock out some steps
that were causing slow downs  Finally, fixed the CUSHAW PE correction
and turned the dataprep ruby script into a shell script.