New Picture, Added HISAT alignment, Readded results collection
New picture showing the purpose and how the Condor python wrapper works.
HISAT alignment option is now available.  Very fast and similar to
Bowtie2 (I have not adjusted the options to make it not gapped, so it
will not be able to utilize RSEM as a counting method).  After a few
changes, we also now can collect results again either in Condor or via
command line.