Integrated UI with GLOW, Changed workflow of UI, Updated RScript with more features
User can now login to GLOW from the UI and request data files (Which can
be selected from the files of a given workflow) and also populate
various other attributes from the database.  Currently it is only
configured for the GLOW test server.

The workflow of the UI has changed to better utilize the expected
workflow.  Thus, fewer tabs are utilized and they are ordered in a
sequential manner.

The RScript has been updated to be able to receive a list of files
loaded from GLOW.