Advanced Aligner and Counter Options
Load advanced aligner and counter options from an XML and present in a
tree view as editable values that can be added to the given aligner or
counter run.  This is currently a WIP.  Untested/Unimplemented within
GLSeq2, though almost implemented at the User Interface level.  Just
need to get the value cells to work out.