Added cleanup script, finalized liblist format, bug fixes
1) Cleanup script now correctly moves files into the final directory for
the user after a run is complete.  Does not yet delete the working
directory (TBD)

2) Liblist format is now finalized in both the R script and the Java GUI

3) Fixed a bug in loading R files where that wasn't working because it
was checking for the UI name and not the attribute file name.  Now first
checks for the hash and then goes through the R file names if it doesn't
find anything (Flexible for the user).