Jobs can now be run from windows
Implemented all GLOW functions and starting jobs (They will auto run via
condor) via SSH.  That means that a user at a windows computer can
easily start GLOW jobs just as if they were at a linux console.  The
login is a bit slow now (It does a verify through both the GLOW database
and the Scarcity servers), so there is room to speed that up a bit.